+++ rust/src/api/mod.rs
@@ -1,1 +1,2 @@
 pub mod device_api;
+pub mod monitor_api;